[General] Use pregenerated bindings for linux
