--- raylib-sys/build.md
@@ -1,1 +1,0 @@
-cmake /Users/david/Documents/rust_stuff/raylib-rs/target/debug/build/raylib-sys-f92607f86a7ad0aa/out/raylib-2.0.0 -DSTATIC=ON -DBUILD_EXAMPLES=OFF -DBUILD_GAMES=OFF -DCMAKE_INSTALL_PREFIX=/Users/david/Documents/rust_stuff/raylib-rs/target/debug/build/raylib-sys-f92607f86a7ad0aa/out -DCMAKE_C_FLAGS="-pthread" -DCMAKE_C_COMPILER=/usr/bin/cc -DCMAKE_CXX_FLAGS=" -ffunction-sections -fdata-sections -fPIC -m64 -pthread" -DCMAKE_CXX_COMPILER=/usr/bin/c++ -DCMAKE_BUILD_TYPE=release
