--- raylib-sys/build.rs.old
@@ -1,35 +1,0 @@
-/* raylib-sys
-   build.rs - Cargo build script
-
-Copyright (c) 2018-2019 Paul Clement (@deltaphc)
-
-This software is provided "as-is", without any express or implied warranty. In no event will the authors be held liable for any damages arising from the use of this software.
-
-Permission is granted to anyone to use this software for any purpose, including commercial applications, and to alter it and redistribute it freely, subject to the following restrictions:
-
-  1. The origin of this software must not be misrepresented; you must not claim that you wrote the original software. If you use this software in a product, an acknowledgment in the product documentation would be appreciated but is not required.
-
-  2. Altered source versions must be plainly marked as such, and must not be misrepresented as being the original software.
-
-  3. This notice may not be removed or altered from any source distribution.
-*/
-
-fn main() {
-    if cfg!(target_os = "windows") {
-        println!("cargo:rustc-link-lib=dylib=gdi32");
-        println!("cargo:rustc-link-lib=dylib=user32");
-    }
-    if cfg!(target_os = "linux") {
-        println!("cargo:rustc-link-search=/usr/local/lib");
-        println!("cargo:rustc-link-lib=X11");
-    }
-    if cfg!(target_os = "macos") {
-        println!("cargo:rustc-link-search=native=/usr/local/lib");
-        println!("cargo:rustc-link-lib=framework=OpenGL");
-        println!("cargo:rustc-link-lib=framework=Cocoa");
-        println!("cargo:rustc-link-lib=framework=IOKit");
-        println!("cargo:rustc-link-lib=framework=CoreFoundation");
-        println!("cargo:rustc-link-lib=framework=CoreVideo");
-    }
-    println!("cargo:rustc-link-lib=static=raylib");
-}
