--- raylib-sys/src/ffi_consts.rs
@@ -1,158 +1,0 @@
-/* automatically generated by rust-bindgen */
-
-pub const PI: f64 = 3.141592653589793;
-pub const DEG2RAD: f64 = 0.017453292519943295;
-pub const RAD2DEG: f64 = 57.29577951308232;
-pub const FLAG_SHOW_LOGO: u32 = 1;
-pub const FLAG_FULLSCREEN_MODE: u32 = 2;
-pub const FLAG_WINDOW_RESIZABLE: u32 = 4;
-pub const FLAG_WINDOW_UNDECORATED: u32 = 8;
-pub const FLAG_WINDOW_TRANSPARENT: u32 = 16;
-pub const FLAG_MSAA_4X_HINT: u32 = 32;
-pub const FLAG_VSYNC_HINT: u32 = 64;
-pub const KEY_SPACE: u32 = 32;
-pub const KEY_ESCAPE: u32 = 256;
-pub const KEY_ENTER: u32 = 257;
-pub const KEY_TAB: u32 = 258;
-pub const KEY_BACKSPACE: u32 = 259;
-pub const KEY_INSERT: u32 = 260;
-pub const KEY_DELETE: u32 = 261;
-pub const KEY_RIGHT: u32 = 262;
-pub const KEY_LEFT: u32 = 263;
-pub const KEY_DOWN: u32 = 264;
-pub const KEY_UP: u32 = 265;
-pub const KEY_PAGE_UP: u32 = 266;
-pub const KEY_PAGE_DOWN: u32 = 267;
-pub const KEY_HOME: u32 = 268;
-pub const KEY_END: u32 = 269;
-pub const KEY_CAPS_LOCK: u32 = 280;
-pub const KEY_SCROLL_LOCK: u32 = 281;
-pub const KEY_NUM_LOCK: u32 = 282;
-pub const KEY_PRINT_SCREEN: u32 = 283;
-pub const KEY_PAUSE: u32 = 284;
-pub const KEY_F1: u32 = 290;
-pub const KEY_F2: u32 = 291;
-pub const KEY_F3: u32 = 292;
-pub const KEY_F4: u32 = 293;
-pub const KEY_F5: u32 = 294;
-pub const KEY_F6: u32 = 295;
-pub const KEY_F7: u32 = 296;
-pub const KEY_F8: u32 = 297;
-pub const KEY_F9: u32 = 298;
-pub const KEY_F10: u32 = 299;
-pub const KEY_F11: u32 = 300;
-pub const KEY_F12: u32 = 301;
-pub const KEY_LEFT_SHIFT: u32 = 340;
-pub const KEY_LEFT_CONTROL: u32 = 341;
-pub const KEY_LEFT_ALT: u32 = 342;
-pub const KEY_RIGHT_SHIFT: u32 = 344;
-pub const KEY_RIGHT_CONTROL: u32 = 345;
-pub const KEY_RIGHT_ALT: u32 = 346;
-pub const KEY_GRAVE: u32 = 96;
-pub const KEY_SLASH: u32 = 47;
-pub const KEY_BACKSLASH: u32 = 92;
-pub const KEY_ZERO: u32 = 48;
-pub const KEY_ONE: u32 = 49;
-pub const KEY_TWO: u32 = 50;
-pub const KEY_THREE: u32 = 51;
-pub const KEY_FOUR: u32 = 52;
-pub const KEY_FIVE: u32 = 53;
-pub const KEY_SIX: u32 = 54;
-pub const KEY_SEVEN: u32 = 55;
-pub const KEY_EIGHT: u32 = 56;
-pub const KEY_NINE: u32 = 57;
-pub const KEY_A: u32 = 65;
-pub const KEY_B: u32 = 66;
-pub const KEY_C: u32 = 67;
-pub const KEY_D: u32 = 68;
-pub const KEY_E: u32 = 69;
-pub const KEY_F: u32 = 70;
-pub const KEY_G: u32 = 71;
-pub const KEY_H: u32 = 72;
-pub const KEY_I: u32 = 73;
-pub const KEY_J: u32 = 74;
-pub const KEY_K: u32 = 75;
-pub const KEY_L: u32 = 76;
-pub const KEY_M: u32 = 77;
-pub const KEY_N: u32 = 78;
-pub const KEY_O: u32 = 79;
-pub const KEY_P: u32 = 80;
-pub const KEY_Q: u32 = 81;
-pub const KEY_R: u32 = 82;
-pub const KEY_S: u32 = 83;
-pub const KEY_T: u32 = 84;
-pub const KEY_U: u32 = 85;
-pub const KEY_V: u32 = 86;
-pub const KEY_W: u32 = 87;
-pub const KEY_X: u32 = 88;
-pub const KEY_Y: u32 = 89;
-pub const KEY_Z: u32 = 90;
-pub const KEY_BACK: u32 = 4;
-pub const KEY_MENU: u32 = 82;
-pub const KEY_VOLUME_UP: u32 = 24;
-pub const KEY_VOLUME_DOWN: u32 = 25;
-pub const MOUSE_LEFT_BUTTON: u32 = 0;
-pub const MOUSE_RIGHT_BUTTON: u32 = 1;
-pub const MOUSE_MIDDLE_BUTTON: u32 = 2;
-pub const MAX_TOUCH_POINTS: u32 = 2;
-pub const GAMEPAD_PLAYER1: u32 = 0;
-pub const GAMEPAD_PLAYER2: u32 = 1;
-pub const GAMEPAD_PLAYER3: u32 = 2;
-pub const GAMEPAD_PLAYER4: u32 = 3;
-pub const GAMEPAD_PS3_BUTTON_TRIANGLE: u32 = 0;
-pub const GAMEPAD_PS3_BUTTON_CIRCLE: u32 = 1;
-pub const GAMEPAD_PS3_BUTTON_CROSS: u32 = 2;
-pub const GAMEPAD_PS3_BUTTON_SQUARE: u32 = 3;
-pub const GAMEPAD_PS3_BUTTON_L1: u32 = 6;
-pub const GAMEPAD_PS3_BUTTON_R1: u32 = 7;
-pub const GAMEPAD_PS3_BUTTON_L2: u32 = 4;
-pub const GAMEPAD_PS3_BUTTON_R2: u32 = 5;
-pub const GAMEPAD_PS3_BUTTON_START: u32 = 8;
-pub const GAMEPAD_PS3_BUTTON_SELECT: u32 = 9;
-pub const GAMEPAD_PS3_BUTTON_UP: u32 = 24;
-pub const GAMEPAD_PS3_BUTTON_RIGHT: u32 = 25;
-pub const GAMEPAD_PS3_BUTTON_DOWN: u32 = 26;
-pub const GAMEPAD_PS3_BUTTON_LEFT: u32 = 27;
-pub const GAMEPAD_PS3_BUTTON_PS: u32 = 12;
-pub const GAMEPAD_PS3_AXIS_LEFT_X: u32 = 0;
-pub const GAMEPAD_PS3_AXIS_LEFT_Y: u32 = 1;
-pub const GAMEPAD_PS3_AXIS_RIGHT_X: u32 = 2;
-pub const GAMEPAD_PS3_AXIS_RIGHT_Y: u32 = 5;
-pub const GAMEPAD_PS3_AXIS_L2: u32 = 3;
-pub const GAMEPAD_PS3_AXIS_R2: u32 = 4;
-pub const GAMEPAD_XBOX_BUTTON_A: u32 = 0;
-pub const GAMEPAD_XBOX_BUTTON_B: u32 = 1;
-pub const GAMEPAD_XBOX_BUTTON_X: u32 = 2;
-pub const GAMEPAD_XBOX_BUTTON_Y: u32 = 3;
-pub const GAMEPAD_XBOX_BUTTON_LB: u32 = 4;
-pub const GAMEPAD_XBOX_BUTTON_RB: u32 = 5;
-pub const GAMEPAD_XBOX_BUTTON_SELECT: u32 = 6;
-pub const GAMEPAD_XBOX_BUTTON_START: u32 = 7;
-pub const GAMEPAD_XBOX_BUTTON_UP: u32 = 10;
-pub const GAMEPAD_XBOX_BUTTON_RIGHT: u32 = 11;
-pub const GAMEPAD_XBOX_BUTTON_DOWN: u32 = 12;
-pub const GAMEPAD_XBOX_BUTTON_LEFT: u32 = 13;
-pub const GAMEPAD_XBOX_BUTTON_HOME: u32 = 8;
-pub const GAMEPAD_ANDROID_DPAD_UP: u32 = 19;
-pub const GAMEPAD_ANDROID_DPAD_DOWN: u32 = 20;
-pub const GAMEPAD_ANDROID_DPAD_LEFT: u32 = 21;
-pub const GAMEPAD_ANDROID_DPAD_RIGHT: u32 = 22;
-pub const GAMEPAD_ANDROID_DPAD_CENTER: u32 = 23;
-pub const GAMEPAD_ANDROID_BUTTON_A: u32 = 96;
-pub const GAMEPAD_ANDROID_BUTTON_B: u32 = 97;
-pub const GAMEPAD_ANDROID_BUTTON_C: u32 = 98;
-pub const GAMEPAD_ANDROID_BUTTON_X: u32 = 99;
-pub const GAMEPAD_ANDROID_BUTTON_Y: u32 = 100;
-pub const GAMEPAD_ANDROID_BUTTON_Z: u32 = 101;
-pub const GAMEPAD_ANDROID_BUTTON_L1: u32 = 102;
-pub const GAMEPAD_ANDROID_BUTTON_R1: u32 = 103;
-pub const GAMEPAD_ANDROID_BUTTON_L2: u32 = 104;
-pub const GAMEPAD_ANDROID_BUTTON_R2: u32 = 105;
-pub const GAMEPAD_XBOX_AXIS_LEFT_X: u32 = 0;
-pub const GAMEPAD_XBOX_AXIS_LEFT_Y: u32 = 1;
-pub const GAMEPAD_XBOX_AXIS_RIGHT_X: u32 = 2;
-pub const GAMEPAD_XBOX_AXIS_RIGHT_Y: u32 = 3;
-pub const GAMEPAD_XBOX_AXIS_LT: u32 = 4;
-pub const GAMEPAD_XBOX_AXIS_RT: u32 = 5;
-pub const MAX_SHADER_LOCATIONS: u32 = 32;
-pub const MAX_MATERIAL_MAPS: u32 = 12;
